Adds new feature for caching dns scenario when upstream goes down
- The new --rcache-preserve-upstream-error command line option
changes the cache behavior so that we only delete expired cache
entries if the upstream is still giving us answers. This provides
a way for systems to continue to work if their primary dns server
goes down. We continue to use the last known result for requests
until the upstream starts to answer queries again. This is off by
default as it's a departure from standard ttl behavior, but can be
a very helpful solution in cloud environments where a system has no
control over the upstream dns server.